create songcard and song data structure
